requirements.txt
certifi==2018.4.16
chardet==3.0.4
future==0.16.0
idna==2.6
python-telegram-bot==10.1.0
requests==2.18.4
urllib3==1.22
virtualenv==16.0.0
